+++ transport-catalogue/transport_catalogue.cpp
@@ -19,7 +19,7 @@ namespace catalogue{
         std::vector<std::string_view> filtered_stops_vector;   
         for (auto& stop : stops){
             if (stops_ptr_.count(stop)){
-                size_t index = std::find(name_stops_.begin(), name_stops_.end(), stop)-name_stops_.begin();
+                size_t index = static_cast<size_t>(std::find(name_stops_.begin(), name_stops_.end(), stop) - name_stops_.begin());
                 filtered_stops_vector.push_back(name_stops_[index]);     
             }    
         }
@@ -31,9 +31,9 @@ namespace catalogue{
         }
     }
 
-    data_bus::InfoBus TransportCatalogue::CountInfoBetweenStations(std::string_view bus ) const{
+    data_bus::InfoBus TransportCatalogue::CountInfoBetweenStations(std::string_view bus ) const {
         int count = 0;
-        int repeat_stop = 0 ;
+        size_t repeat_stop = 0 ;
         std::map<std::string_view, size_t> statistics_repeat;
         Coordinates from;
         Coordinates to;
@@ -41,7 +41,7 @@ namespace catalogue{
 
         std::vector<std::string_view> rezult = buses_.at(bus).MakeFullBus();
         double distance_track = CountDist(rezult);
-        int count_stops = rezult.size();
+        size_t count_stops = rezult.size();
         for (auto stop : rezult ){
             ++statistics_repeat[stop];
             to=stops_ptr_.at(stop)->point;    
@@ -55,8 +55,6 @@ namespace catalogue{
             from = to;
             count++;
         }
-        
-    
         double curvature = distance_track/distance_geo;
         return {buses_.at(bus).name, count_stops, count_stops-repeat_stop, distance_geo, distance_track,
                 curvature, buses_.at(bus).is_roundtrip};
@@ -69,22 +67,22 @@ namespace catalogue{
         return CountInfoBetweenStations(id);
     }
 
-    data_bus::InfoStop TransportCatalogue::FindStop(std::string_view id) const{
+    data_bus::InfoStop TransportCatalogue::FindStop(std::string_view id) const {
         if(!stops_ptr_.count(id)){
             return data_bus::InfoStop();
         }
         return ReturnBusesWithStop(id);
     }
 
-    const std::unordered_map<std::string_view, data_bus::Bus>& TransportCatalogue::GetAllBus() const{
+    const std::unordered_map<std::string_view, data_bus::Bus>& TransportCatalogue::GetAllBus() const {
         return buses_;
     }
 
-    const std::unordered_map<std::string_view, data_bus::Stop *>& TransportCatalogue::GetAllstops() const{
+    const std::unordered_map<std::string_view, data_bus::Stop *>& TransportCatalogue::GetAllstops() const {
         return stops_ptr_;
     }
 
-    data_bus::InfoStop TransportCatalogue::ReturnBusesWithStop(std::string_view id ) const{
+    data_bus::InfoStop TransportCatalogue::ReturnBusesWithStop(std::string_view id ) const {
         data_bus::InfoStop collect;
         collect.name = stops_ptr_.at(id)->name;
         if(!stops_to_buses_.count(id)){
@@ -96,7 +94,7 @@ namespace catalogue{
         return collect;
     }
 
-    double TransportCatalogue::CountDist(std::vector<std::string_view> list_stop) const{
+    double TransportCatalogue::CountDist(std::vector<std::string_view> list_stop) const {
         double count = 0;
         for (size_t i = 0; i < list_stop.size() - 1; i++){//идем по остановкам   
             auto straight = ReturnStopsDistance(list_stop.at(i), list_stop.at(i + 1));//создадим пару так как она есть
@@ -111,9 +109,9 @@ namespace catalogue{
     }
 
     void TransportCatalogue::AddStopsDistance(std::string_view stop_one, std::string_view stop_two, int distance){
-        size_t one = std::find(name_stops_.begin(), name_stops_.end(), stop_one)-name_stops_.begin();
-        size_t two = std::find(name_stops_.begin(), name_stops_.end(), stop_two)-name_stops_.begin();
-        if ( two>=name_stops_.size() ){
+        size_t one = static_cast<size_t>(std::find(name_stops_.begin(), name_stops_.end(), stop_one)-name_stops_.begin());
+        size_t two = static_cast<size_t>(std::find(name_stops_.begin(), name_stops_.end(), stop_two)-name_stops_.begin());
+        if (two>=name_stops_.size()){
             name_stops_.push_back(std::string(stop_two));
             std::pair<std::string_view,std::string_view> dist(name_stops_[one],name_stops_.back());
             stop_distance_[dist] = distance;
